changed to hex output
git-svn-id: file:///srv/svn/spinny@415 9787d06c-1a0a-0410-a88a-902e1a4f81df

+++ libs/ews/connection.cpp
@@ -57,7 +57,7 @@ namespace ews {
 				request_, buffer_.data(), buffer_.data() + bytes_transferred );
 			
 			if (result) {
-				BOOST_LOGL(www,debug) << "Begin Write: " <<  (int)this
+				BOOST_LOGL(www,debug) << "Begin Write: " <<  this
 							<< " : " << request_.uri << std::endl;
  	
 				if ( ! request_handler::handle_request( request_, reply_ ) ){
@@ -100,7 +100,7 @@ namespace ews {
 		}
 		BOOST_LOGL(www,debug)
 			<< "Wrote " << bytes_transferred << " bytes on connection "
-			<< (int)this << " result: " << e.what();
+			<< this << " result: " << e.what();
 	}
 
 } // namespace ews
